requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml --resolver=backtracking --all-extras -o requirements.txt
annotated-types==0.7.0
    # via pydantic
attrs==24.2.0
    # via
    #   frictionless
    #   jsonlines
    #   jsonschema
    #   referencing
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
cfgv==3.4.0
    # via pre-commit
chardet==5.2.0
    # via frictionless
charset-normalizer==3.4.0
    # via requests
ckanapi==4.8
    # via hdx-python-api
click==8.1.7
    # via typer
coverage==7.6.4
    # via pytest-cov
defopt==6.4.0
    # via hdx-python-api
distlib==0.3.9
    # via virtualenv
dnspython==2.7.0
    # via email-validator
docopt==0.6.2
    # via
    #   ckanapi
    #   num2words
docutils==0.21.2
    # via defopt
email-validator==2.2.0
    # via hdx-python-api
et-xmlfile==2.0.0
    # via openpyxl
filelock==3.16.1
    # via virtualenv
frictionless==5.18.0
    # via hdx-python-utilities
google-auth==2.36.0
    # via
    #   google-auth-oauthlib
    #   gspread
google-auth-oauthlib==1.2.1
    # via gspread
gspread==6.1.4
    # via hdx-python-scraper
hdx-python-api==6.3.4
    # via
    #   hdx-scraper-ophi (pyproject.toml)
    #   hdx-python-scraper
hdx-python-country==3.8.3
    # via
    #   hdx-scraper-ophi (pyproject.toml)
    #   hdx-python-api
    #   hdx-python-scraper
hdx-python-scraper==2.5.0
    # via hdx-scraper-ophi (pyproject.toml)
hdx-python-utilities==3.7.4
    # via
    #   hdx-scraper-ophi (pyproject.toml)
    #   hdx-python-api
    #   hdx-python-country
    #   hdx-python-scraper
humanize==4.11.0
    # via frictionless
identify==2.6.1
    # via pre-commit
idna==3.10
    # via
    #   email-validator
    #   requests
ijson==3.3.0
    # via hdx-python-utilities
inflect==7.4.0
    # via quantulum3
iniconfig==2.0.0
    # via pytest
isodate==0.7.2
    # via frictionless
jinja2==3.1.4
    # via frictionless
jsonlines==4.0.0
    # via hdx-python-utilities
jsonpath-ng==1.7.0
    # via libhxl
jsonschema==4.23.0
    # via
    #   frictionless
    #   tableschema-to-template
jsonschema-specifications==2024.10.1
    # via jsonschema
libhxl==5.2.2
    # via
    #   hdx-python-api
    #   hdx-python-country
loguru==0.7.2
    # via hdx-python-utilities
makefun==1.15.6
    # via hdx-python-api
markdown-it-py==3.0.0
    # via rich
marko==2.1.2
    # via frictionless
markupsafe==3.0.2
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
more-itertools==10.5.0
    # via inflect
nodeenv==1.9.1
    # via pre-commit
num2words==0.5.13
    # via quantulum3
oauthlib==3.2.2
    # via requests-oauthlib
openpyxl==3.1.5
    # via hdx-python-utilities
packaging==24.1
    # via pytest
petl==1.7.15
    # via frictionless
platformdirs==4.3.6
    # via virtualenv
pluggy==1.5.0
    # via pytest
ply==3.11
    # via
    #   jsonpath-ng
    #   libhxl
pockets==0.9.1
    # via sphinxcontrib-napoleon
pre-commit==4.0.1
    # via hdx-scraper-ophi (pyproject.toml)
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pydantic==2.9.2
    # via frictionless
pydantic-core==2.23.4
    # via pydantic
pygments==2.18.0
    # via rich
pyphonetics==0.5.3
    # via hdx-python-country
pytest==8.3.3
    # via
    #   hdx-scraper-ophi (pyproject.toml)
    #   pytest-check
    #   pytest-cov
pytest-check==2.4.1
    # via hdx-scraper-ophi (pyproject.toml)
pytest-cov==6.0.0
    # via hdx-scraper-ophi (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   frictionless
    #   hdx-python-utilities
    #   libhxl
python-io-wrapper==0.3.1
    # via libhxl
python-slugify==8.0.4
    # via
    #   ckanapi
    #   frictionless
pyyaml==6.0.2
    # via
    #   frictionless
    #   pre-commit
    #   tableschema-to-template
quantulum3==0.9.2
    # via hdx-python-api
ratelimit==2.2.1
    # via hdx-python-utilities
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via hdx-python-scraper
requests==2.32.3
    # via
    #   ckanapi
    #   frictionless
    #   hdx-python-api
    #   libhxl
    #   requests-file
    #   requests-oauthlib
requests-file==2.1.0
    # via hdx-python-utilities
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rfc3986==2.0.0
    # via frictionless
rich==13.9.4
    # via typer
rpds-py==0.21.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
ruamel-yaml==0.18.6
    # via hdx-python-utilities
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
setuptools==75.3.0
    # via ckanapi
shellingham==1.5.4
    # via typer
simpleeval==1.0.3
    # via frictionless
simplejson==3.19.3
    # via ckanapi
six==1.16.0
    # via
    #   ckanapi
    #   pockets
    #   python-dateutil
    #   sphinxcontrib-napoleon
sphinxcontrib-napoleon==0.7
    # via defopt
stringcase==1.2.0
    # via frictionless
structlog==24.4.0
    # via libhxl
tableschema-to-template==0.0.13
    # via hdx-python-utilities
tabulate==0.9.0
    # via frictionless
tenacity==9.0.0
    # via hdx-python-country
text-unidecode==1.3
    # via python-slugify
typeguard==4.4.1
    # via inflect
typer==0.13.0
    # via frictionless
typing-extensions==4.12.2
    # via
    #   frictionless
    #   pydantic
    #   pydantic-core
    #   typeguard
    #   typer
unidecode==1.3.8
    # via
    #   libhxl
    #   pyphonetics
urllib3==2.2.3
    # via
    #   libhxl
    #   requests
validators==0.34.0
    # via frictionless
virtualenv==20.27.1
    # via pre-commit
wheel==0.44.0
    # via libhxl
xlrd==2.0.1
    # via hdx-python-utilities
xlrd3==1.1.0
    # via libhxl
xlsx2csv==0.8.3
    # via hdx-python-utilities
xlsxwriter==3.2.0
    # via tableschema-to-template
xlwt==1.3.0
    # via hdx-python-utilities
